requirements.txt
gunicorn
django-heroku
lxml
djangorestframework
requests
cssselect